+++ src/Product/database/migrations/Denma_01_03_002_Product.php
@@ -19,8 +19,8 @@ class Product extends Migration
             $table->integer('product_group_id')->unsigned()->nullable()->default(null);
             $table->float('min_order');
             $table->boolean('customizable')->default(true);
-            $table->float('base_price')->default(0);
-            $table->float('per_unit_price')->default(0);
+            $table->bigInteger('base_price')->default(0);
+            $table->bigInteger('per_unit_price')->default(0);
             $table->enum('status', ['draft', 'active', 'inactive'])->default('draft');
             $table->timestamps();
             $table->softDeletes();
